+++ python/slack_types/web_api/rtm_start_response.py
@@ -2076,10 +2076,10 @@ class Message1(BaseModel):
 class Field1(BaseModel):
     key: str | None = None
     column_id: str | None = None
-    value: str | None = None
+    value: bool | int | str | None = None
     text: str | None = None
     rich_text: List[Any] | None = None
-    message: Message1 | None = None
+    message: List[Message1] | Message1 | None = None
     number: List[Any] | None = None
     select: List[Any] | None = None
     date: List[Any] | None = None
@@ -2521,10 +2521,10 @@ class Message2(BaseModel):
 class Field2(BaseModel):
     key: str | None = None
     column_id: str | None = None
-    value: str | None = None
+    value: bool | int | str | None = None
     text: str | None = None
     rich_text: List[Any] | None = None
-    message: Message2 | None = None
+    message: List[Message2] | Message2 | None = None
     number: List[Any] | None = None
     select: List[Any] | None = None
     date: List[Any] | None = None
@@ -5401,10 +5401,10 @@ class Message4(BaseModel):
 class Field5(BaseModel):
     key: str | None = None
     column_id: str | None = None
-    value: str | None = None
+    value: bool | int | str | None = None
     text: str | None = None
     rich_text: List[Any] | None = None
-    message: Message4 | None = None
+    message: List[Message4] | Message4 | None = None
     number: List[Any] | None = None
     select: List[Any] | None = None
     date: List[Any] | None = None
@@ -5831,10 +5831,10 @@ class Message5(BaseModel):
 class Field6(BaseModel):
     key: str | None = None
     column_id: str | None = None
-    value: str | None = None
+    value: bool | int | str | None = None
     text: str | None = None
     rich_text: List[Any] | None = None
-    message: Message5 | None = None
+    message: List[Message5] | Message5 | None = None
     number: List[Any] | None = None
     select: List[Any] | None = None
     date: List[Any] | None = None